Rewrite of io_init()
The introduction of the "-C <confdir>" option is an opportunity to review the
initialization of data paths. It lead to a rewrite.

Two "root" directories are used (data and configuration files); by default they
are identical. The statically allocated path buffers are protected against
overflow. Buffer allocation is moved to vars.c.

Missing files/directories now include hooks.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>