DST fix: daylength v. DAYINSEC
The number of seconds in a day and daylength in seconds differ when
Daylight Saving Time is in effect on two days of the year. The day when DST
takes effect is 23 hours long, and the day when DST ends is 25 hours long.

In the latter case the date changing thread wóuld enter a loop in the last hour
before midnight (in the former it would set the date an hour too late).
The next midnight is calculated through mktime(), invoked by date2sec().

Wrong daylength prevented appointments from being stored in the day vector and
caused them to be displayed wrongly in the appts panel.