Fix day range for queries.
In "--from a --to z", a is included in the range, z not. This is non-intuitive
and disagrees with the semantics of "to" in filter options like --filter-start-to,
where "to" (and "from") are used inclusively (as opposed to "before" and
"after"). It also has the effect that "--from today --to tomorrow" has a range
of 1 day, "--to z" a range of 0 days (otherwise not allowed), and "--to today
--days -1" is allowed and displays yesterday!

The implementation has been fixed to agree with "inclusive" semantics. Options
--from and -days with negative range are allowed, while --to and --days are
disallowed also when the range is negative.