CLI: filter options have no effect for dates before the epoch
With the exception of filter.type_mask, a filter is only applied if set
explicitly on the command line with a filter option. Whether that is the case,
is determined by comparison with the initialization value. For date related
filters (start_from/to, end_from/to) that is -1, hence the criterion is != -1,
not >= 0.

In generel, a filter initialization value should be invalid (i.e. one that
cannot be set explicitly). As times before the epoch (1 January 1970 00:00:00
UTC) are negative, -1 is a valid Unix time. However, as it cannot be set from
the command line, it is probably no problem?

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>