Safety exit and read-only mode
The key_generic_command() function provides a "safety exit" in case of online
changes that should be dropped at exit when auto_save is on, or (NEW) should be
saved at exit when in read-only mode. A check for unsaved changes has been
added. The command prompt has been extended with minimal help information.

After commit 05e0fd0 "Quit, autosave and interactive save" you could not leave
calcurse in read-only mode with the quit command because the key_generic_quit()
function interpreted IO_SAVE_CANCEL as a decision (by the user) to cancel a
save-conflict and cancelled the quit as well. Now the function will quit
unconditionally in read-only mode (as stated in the man page). In normal mode a
check for unsaved changes has been added in case auto-save is off.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>