Filter option: invert
New filter option: --filter-invert. When present it inverts (negates) the other
filter options combined. This is mostly useful with the -G option (with -Q the
output is limited by the query range (day range)).

The ouput from "calcurse -G <filter options>" is the (set) complement of "calcurse
-G <filter options> --filter-invert". Here <filter options> may be any combination
of filter options.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>