Purge option glorified
The invert filter option is automatically set for -P (and cannot be used on the
command line). The intention is that the grep option (-G) is used to find the
items that should be removed. To remove them, the same command is run with -P
instead of -G.

In general, purge will remove matching items (silently).

Backward compatibility for option -F (no invert filter)..

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>