Simplifications in day storage
The function day_process_storage() is a wrapper for day_store_items(). It has an
unused second argument, and is only used twice to load the selected day. It has
been removed.

A new function, get_slctd_day(), is the equivalant of get_today() and replaces
the very awkwardly named ui_calendar_get_slctd_day_sec().