Redesign selected-item implementation for the APP panel
The day vector, day_items, is displayed in the appointments panel; the selected
day_item object is highlighted (when the panel has the focus). When items are
inserted, edited, moved etc., and when the day is changed, the day vector is
rebuilt and displayed anew.

Problem: How shall the selection be set automatically in the context of the new
day vector?

In previous versions all of the above is mostly handled by the function
do_storage() in calcurse.c The function saves data about the selection as
needed, rebuilds the day vector, loads the listbox and sets the selection from
the saved selection data.  This works well in "single day" calcurse in cases
where the selected item is present in the day vector both before and after the
rebuild, or when the item ordering in the listbox is unaffected by the changes.
But when a new item is added the selection cannot be set to the new object by
do_storage().  Instead the necessary operations are performed by
ui_day_item_add(), and do_storage() is bypassed. In general, when an item cannot
be found in the new vector, the item which occupies the old place in the list
gets selected, e.g. when an item is deleted. When an item is turned into a
repeating one, the old item is deleted and a new is created. Here the new
selection is not always the affected item, but in any case not far away.
Generally, with only one day in the panel an erronous selection might not be
noticed or be accurate by chance.

In "multiple day" calcurse the existing scheme works less well; in addition the
day vector may now contain more than one object that refer to the same event or
appointment (recurrent items or multi-day appointments). The scheme has
therefore been modified. The do_storage() function is no longer bypassed, but
handles day vector rebuild, load of listbox and item selection exclusively. To
make that possible, data about the selected item is no longer saved in a local
automatic variable, private to do_storage(), but in an external static variable
in day.c, which may be set not only by do_storage(). The variable is declared as

static struct day_item sel_data;

and used as follows:

1. On startup sel_data is initialized to empty (i.e. no selection).
2. In any operation involving the appointments panel:
	2.1 Do the work and if necessary set sel_data. This is the case when
	    deleting, adding or pasting an item, and when turning an ordinary
	    item into a recurrent one.
	2.2 Call do_storage().
3. In do_storage():
	3.1 If sel_data is empty, set it to the current selection.
	3.2 Rebuild the day vector.
	3.3 Set the selection from sel_data.
	3.4 Set sel_data to empty.

Further remarks

The selection is found in the new day vector by searching for the saved (order,
item.<pointer>) pair. Previously the item.<pointer> alone sufficed and in some
cases it still does. In case the item cannot be found, the selection stays in
the same day as before the rebuild.

An attempt at more consistently named APP-related functions has led to:
ui_day_sel_date() replaces ui_day_sel_day()
ui_day_get_sel() replaces ui_day_selitem()

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>