Redesign of selected-item implementation for the APP panel
The day vector, day_items, is displayed in the APP listbox; the selected
day_item object is highlighted. The problem is, how to keep track of the
selection and set it correctly when day_item objects are inserted, edited, moved
etc., and when the day is changed.

The function do_storage() in calcurse.c rebuilds the day_items vector
repeatedly, so pointers into the vector must be used carefully. Instead an
external static variable, struct day_item slctd_item, saves data from the which
the selection can be set after the rebuild. The variable has file scope only,
but may be set externally with cal_set_sel().

Usage:
1. Initialize slctd_item to empty (i.e. no selection).
2. In any operation involving the APP panel:
	2.1 Do the necessary work.
	2.1 If the selection must be set explicitly, call cal_set_sel().
	    This is the case when adding or pasting an item, and when turning an
	    ordinary item into a recurrent item.
	2.2 Call do_storage().
3. In do_storage():
	3.1 If slctd_item is empty, save the current selection.
	3.2 Rebuild the day vector.
	3.3 Set the selection from slctd_item.
	3.4 Set slctd_item to empty.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>