APP panel scrolling
With multiple days in the APP panel, up/down movements should change behaviour
at the top and bottom of the list displayed, and load the previous/next lot of
days.

This requires that the move function returns the result of the operation.
Furthermore, the ability to move the selection to the beginning of a day is
needed when moving down (in order to move from the first day to the last day).
For this reason a DAY_SEPARATOR has been inserted also after the last day of a
lot.

The function day_get_nb() has been renamed day_get_days().

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>