Multiple days: appointment height
Appointments have a listbox height of three to separate them clearly when there
is more than one in a day.  This leaves a spurious empty line at the end of a
day with appointments.

The height is changed to two, and an APP_SEPARATOR is inserted. After sorting of
the listbox contents superfluous separators are removed.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>