Multiple days: calendar display
Movements in the calendar is based on the selected day, struct date slctd_day,
in ui-calendar.c. On the screen it stands out with a deviating colour.

The scheme is extended to multiple days in the APP panel as follows. Instead of
a colour, a pair of square brackets is used, which fences in the days displayed
in the APP panel. An alternative "bar" is also implemented. The selected day is
the first day of the range.

At this point the bracket fence, like the number of days (five), is fixed
(hard-coded). Both may be made configurable.

The drawing function draw_monthly_view() has been cleaned up and made easier to
read, but is basically unchanged apart from the selection fence/bar.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>