Add week numbers in the calendar and full first and last week
Much in the calendar is based on the selected day, struct date slctd_day,
in ui-calendar.c.

On the screen it is highlighted with a deviating colour. The highlight effect
has been changed to a pair of red square brackets that do not obscure the day
colour.

The week number (in the frame) used to be that of the selected day, but has
no obvious relation to the days in the APP panel. It has been replaced by the
year day number of the selected day. The week numbers of all visible weeks are
displayed to the left of the calendar.

Dates are displayed also for the overlapping parts of the first and last
week of the month (which do not belong to the month).

Days are accessible in the appoinments panel as well as in the calendar. Hence,
validation of days (= inside UNIX time limits) must be extended from the calendar
(in ui_calendar_move()) to include loaded days (in day_store_items()).

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>