Add configuration variables for multiple days
The number of days displayed in the APP panel has been made configurable,
maximum 21 days, default seven days.

With several days in the APP panel, it may be desirable to "squeeze" the entries
by leaving out the final empty line of each appointment and lower the number of
lines between consecutive days (0, 1, or 2). Both are made general configuration
options. To make a uniform display, an empty line is added to a day without
appointments, if appointments have an empty line.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>