Multiple days: configuration variables
The number of days displayed in the APP panel (1,..,7) and their indication in
the CAL panel (red brackets or bar) have been made configurable.

With the number of days set to one, the look-and-feel is that of the 'old'
calcurse with the addition of arrow-key movements in the APP panel.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>