Scrollwindow pad improvement
A scrollwindow consists of a pad to write on, and a window through which to view
the pad, the viewport. The pad may change in size when new contents are loaded
into the scrollwindow. If so, the pad size is set with
wins_scrollwin_set_linecount(). But the offset into the pad (the top line to be
displayed in the viewport) is not adjusted, although it may not be appropriate
for the new pad size. This is probably the cause of the problem solved by
commit 0b46ad4, Avoid blank space after the last list box item,

The wins_scrollwin_set_linecount() has been renamed wins_scrollwin_set_pad() and
sets size as well as offset. The offset is only changed if necessary.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>