Explicit calendar and todo view configuration
Problem
-------
The configuration settings for calendar view (monthly/weekly) and todo view
(hide/show completed) used to be saved automatically on calcurse exit, with
values taken from the current interactive settings. They could not be set
explicitly in the configuration menues. Configuration settings are no longer
saved on program exit, but on exit from the configuration menu. This means that
the saved values are those that were current when the configuration menu was
entered. To change a saved value, you must set the view as desired and then
enter/exit the configuration menu.

Solution
--------
The preferred calendar and todo views are no longer automatically taken from the
interactive settings, but are explicitly set in the general options menu.
Default values are monthly view and hide completed view.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>