Multiple days: fix appointment and day separation
Previously appointments had a height of three with text only in two lines. This
created an empty line between successive appointments.

With multiple days in the APP panel, days are separated by a DAY_SEPARATOR of
height two. A day without an appointment thus ends in two empty lines, while a
day with an appointment ends in three empty lines.

Two lines seem an appropriate separation between days. To avoid the extra line
after the last appointment, the appointment height was reduced to two and an
APPT_SEPARATOR inserted after each; the APPT_SEPARATOR got an order one second
greater than the appointment.  The last APPT_SEPARATOR of each day, if any, was
removed after sorting of the day vector.

The scheme breaks down when two appointments on the same day have the same start
time. The two APPT_SEPARATORS are then sorted after the appointments. The scheme
is furthermore not very general and solves the problem in too many places.

-------------

New scheme. The appointment height is set to three (again), and the
APPT_SEPARATOR is eleminated. The DAY_SEPARATOR height is reduced from two to
one, and a new EMPTY_SEPARATOR of height one is inserted in any day with only
events.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>