Multiple days: selected day derived from selected item
Previously, with only one day visible at a time, the appointments panel
displayed the details of the day selected in the calendar (slctd_day);
information required for operations on items (day_items) can often be derived
from the selected day. The items available are derived from the selected day. In
particular, the selected item is derived from the selected day.

With multiple days in the APP panel, the relation between selected day (in the
calendar) and the selected item (in the APP panel) has, in a way, been turned
around. The selected item may now be moved between days without explicitly
changing the selected day. Implicitly it is changed when the target day of a
move is unavailable.

This commit draws the full consequence: the selected day in the calendar is
always (set to) the day of the selected item in the APP panel.

The static variable 'struct date slctd_day' lives in ui_calendar.c and is
accessible through various public functions. To these are added
ui_calendar_set_slctd_day() which sets slctd_day directly.

In the calendar the selected day is marked by a pair of highlighted square
brackets.

The selected day retains its significance for load of the day vector (in
day_store_items()): the range of loaded days begins with the selected day.
Movements (up/down) in the APP panel will change the selected day as the
selected item moves among the already loaded days. Only when the target of a
movement is unreachable, will further days be loaded. On the other hand, if the
same range of days must be reloaded because of a changed item, the selected item
- and with it the selected day - must be reset to the first day item (see
do_storage()).

Movements in the calendar (generic-next-day, etc.) are not affected and behave
as previously, i.e. they will cause a range of days to be loaded with the
selected day as the first and the selected item as the first of the selected
day.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>