View or edit exception days of a recurrent item
The exception days are presented for viewing/editing as a string of
space-separated dates (in the user-preferred input format). After editing the
string is checked for valid dates, but there is no check that a date is
meaningful (an occurrence day of the item between start day and until day).
Although possible, it is best to add exception days in the usual way by deletion
of occurrences.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>