Fix initialization of the day_items vector
This patch fixes all failings tests, but one, in PR #193.

Until now the missing initialization of day_items_nb has caused no problems,
because the variable was assigned to (=) before being used. In the Multiple days
implementation it is repeatedly increased (+=) in a loop without being
initialized first. Indeed, this may considered an easily fixed bug. But the
initialization really belongs in day_init_vector() so that the call
day_item_count(0) returns 0 if done right after the call day_init_vector(). The
bug only shows up in command line mode because day_items_nb is not used in
interactive mode.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>