Implement scrolling in the appointments panel
With multiple days in the APP panel, up/down movements should change behaviour
at the top and bottom of the list displayed, and load the previous/next lot of
days.

This requires that the move function returns the result of the operation.
Furthermore, the ability to move the selection to the beginning of a day is
needed when moving down (in order to move from the first day to the last day).
For this reason a DAY_SEPARATOR has been inserted also after the last day of a
lot.

Appointments have a listbox height of three to separate them clearly when there
is more than one in a day.  This leaves a spurious empty line at the end of a
day with appointments.  The DAY_SEPARATOR height is reduced from two to one, and
a new EMPTY_SEPARATOR of height one is inserted in any day with only events.

When scrolling up the DAY_HEADING becomes visible when the selection reaches the
first item of the day.

The length of the separator (between events and appointmets) is adjusted to
leave a space to the window border at both ends, thereby making it a part of the
day, not a separation between days.

The dummy event must also be recognisable when not the selected item and is only
inserted in interactive mode.

The test for a saved selection must also recognise caption items which have item
pointer NULL.

The function day_get_nb() has been renamed day_get_days().

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>