Fix display of time left before next appointment
In the notify bar, the clock is shown in hours, minutes and seconds, whereas the
time left to the next appointment is shown in hours and minutes only. When you
read the clock in hours and minutes (discarding the seconds), you are mentally
rounding down to the nearest minute. To agree with that reading, the time left
(in seconds) should be (programmatically) rounded up to the nearest minute for
display.

In that way the time left is counted down whenever the minute "hand" of the
clock "ticks", and reaches zero when the clock reaches the time of the next
appointment, not one minute before.  Also, the clock (in hours and minutes)and
the time left always add up to the time of the next appointment.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>