Refactor function calls: recurrence parameters as a single argument
Background. The recurrence parameters are type, frequency, until date and
exception list (in RFC 5545 parlance FREQ, INTERVAL, UNTIL and EXDATE's). When
these are passed in a function call, the argument list becomes long and not very
readable.  When support for extended recurrence rules is implemented, the number
of recurrence parameters increases, and function signatures must be amended.

Solution. The "struct rpt" is extended with the exception list; any future
recurrence parameters are added here. A pointer to this structure replaces the
recurrence parameters in function calls.

Note. Each recurrent event and appoinment instance has (a pointer to) a "struct
rpt" and in addition an exception list. The latter is retained to avoid the
derived changes, and the exception list in the structure is initialized to an
empty list when the recurrent instance is created.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>