Set time-of-day consistently to midnight for until day
A day begins on midnight (inclusive) and ends on midnight (exclusive). A day
as a whole is represented by the initial midnight, i.e. time-of-day is 00:00.

On load of recurrent appointments (but not events) time-of-day for the until day
is set to 23:59. For a newly created recurrent appointment the setting depends
on the input method: time-of-day is set to 00:00 if until day is given as a date
(day, month and year), but to time-of-day for the start day if given as an
offset (+dd).

The resulting behaviour is only visible in interactive use of calcurse as proved
by the following scenario.

1) Create an appointment with start time 12:00, end time 11:59 (multi day)

2) Turn it into a recurrent appointment of type daily, frequency 3, until day +3

The appointment is correctly displayed with two 2-day occurrences three days
apart.

3) Edit the appointment and select Repetition. Accept existing type, frequency
and end day (now as a date)

The second day of the second occurrence disappears.

4) Repeat 3), but set the end day as an offset (+3)

The second day of the second occurrence reappears.

The inconsistencies have been eliminated, and time-of-day for the until day is
now always 00:00.

Also, until day may equal start day, so midnights should be compared.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>